LayoutTests/http/wpt/loading/early-hints/preconnect.h2.window-expected.txt

CONSOLE MESSAGE: Preconnecting to https://127.0.0.1:9000/ due to early hint


PASS Early hints preconnect
